+++ examples/basic/src/model-provider/mistral/mistral-chat-stream-text-example.ts
@@ -5,14 +5,16 @@ dotenv.config();
 
 async function main() {
   const textStream = await streamText(
-    mistral
-      .TextGenerator({
-        model: "mistral-medium",
-        maxGenerationTokens: 120,
-      })
-      .withTextPrompt(),
-
-    "Write a short story about a robot learning to love:"
+    mistral.ChatTextGenerator({
+      model: "mistral-medium",
+      maxGenerationTokens: 120,
+    }),
+    [
+      {
+        role: "user",
+        content: "Write a short story about a robot learning to love:",
+      },
+    ]
   );
 
   for await (const textPart of textStream) {