+++ src/model-provider/ollama/OllamaFacade.ts
@@ -7,6 +7,14 @@ import {
   OllamaCompletionModel,
   OllamaCompletionModelSettings,
 } from "./OllamaCompletionModel.js";
+import {
+  OllamaApiConfiguration,
+  OllamaApiConfigurationSettings,
+} from "./OllamaApiConfiguration.js";
+
+export function Api(settings: OllamaApiConfigurationSettings) {
+  return new OllamaApiConfiguration(settings);
+}
 
 export function CompletionTextGenerator<CONTEXT_WINDOW_SIZE extends number>(
   settings: OllamaCompletionModelSettings<CONTEXT_WINDOW_SIZE>